add validation

+++ api/src/main/java/bm/service/impl/BookServiceImpl.java
@@ -1,13 +1,17 @@
 package bm.service.impl;
 
+import bm.entity.dto.book.BookAddingDTO;
+import bm.entity.dto.book.BookUpdatingDTO;
 import bm.mapper.BookMapper;
-import bm.model.Book;
+import bm.entity.model.Book;
 import bm.service.BookService;
+import cn.hutool.core.bean.BeanUtil;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @Author: Gorgeous
@@ -23,7 +27,14 @@ public class BookServiceImpl extends ServiceImpl<BookMapper, Book> implements Bo
 
     @Override
     public Book retrieveBookById( Long id ) {
-        return bookMapper.selectById(id);
+        if(Objects.isNull(id)){
+            throw new RuntimeException("id can not be empty");
+        }
+        Book book = bookMapper.selectById(id);
+        if(Objects.isNull(book)){
+            throw new RuntimeException("no book id such like "+ id);
+        }
+        return book;
     }
 
     @Override
@@ -32,17 +43,22 @@ public class BookServiceImpl extends ServiceImpl<BookMapper, Book> implements Bo
     }
 
     @Override
-    public boolean addBook( Book book ) {
+    public boolean addBook( BookAddingDTO bookAddingDTO ) {
+        Book book = BeanUtil.toBean(bookAddingDTO, Book.class);
         return bookMapper.insert(book) > 0;
     }
 
     @Override
     public boolean deleteBook( Long id ) {
+        if(Objects.isNull(id)){
+            throw new RuntimeException("id can not be empty");
+        }
         return bookMapper.deleteById(id) > 0;
     }
 
     @Override
-    public int updateBook( Book book ) {
+    public int updateBook( BookUpdatingDTO bookUpdatingDTO ) {
+        Book book = BeanUtil.copyProperties(bookUpdatingDTO, Book.class);
         return bookMapper.updateById(book);
     }
